+++ src/ansys/aedt/core/circuit_netlist.py
@@ -28,8 +28,8 @@ from pathlib import Path
 import shutil
 
 from ansys.aedt.core.application.analysis_circuit_netlist import AnalysisCircuitNetlist
-from ansys.aedt.core.generic.filesystem import search_files
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
+from ansys.aedt.core.internal.filesystem import search_files
 
 
 class CircuitNetlist(AnalysisCircuitNetlist, object):
