+++ src/ansys/aedt/core/visualization/advanced/misc.py
@@ -32,9 +32,9 @@ from ansys.aedt.core.generic.constants import CSS4_COLORS
 from ansys.aedt.core.generic.constants import SI_UNITS
 from ansys.aedt.core.generic.constants import unit_system
 from ansys.aedt.core.generic.file_utils import open_file
-from ansys.aedt.core.generic.filesystem import search_files
 from ansys.aedt.core.generic.general_methods import pyaedt_function_handler
 from ansys.aedt.core.generic.load_aedt_file import load_keyword_in_aedt_file
+from ansys.aedt.core.internal.filesystem import search_files
 from ansys.aedt.core.modeler.geometry_operators import GeometryOperators
 
 try:
