+++ src/ansys/aedt/core/workflows/project/kernel_converter.py
@@ -34,9 +34,9 @@ from ansys.aedt.core import Maxwell3d
 from ansys.aedt.core import Q3d
 from ansys.aedt.core.application.design_solutions import solutions_types
 from ansys.aedt.core.generic.file_utils import generate_unique_name
-from ansys.aedt.core.generic.filesystem import search_files
 from ansys.aedt.core.generic.settings import settings
 from ansys.aedt.core.internal.design_types import get_pyaedt_app
+from ansys.aedt.core.internal.filesystem import search_files
 import ansys.aedt.core.workflows
 from ansys.aedt.core.workflows.misc import get_aedt_version
 from ansys.aedt.core.workflows.misc import get_arguments
