+++ tests/system/solvers/conftest.py
@@ -71,7 +71,7 @@ from ansys.aedt.core.aedt_logger import pyaedt_logger
 from ansys.aedt.core.filtersolutions import DistributedDesign
 from ansys.aedt.core.filtersolutions import LumpedDesign
 from ansys.aedt.core.generic.file_utils import generate_unique_name
-from ansys.aedt.core.generic.filesystem import Scratch
+from ansys.aedt.core.internal.filesystem import Scratch
 
 local_path = os.path.dirname(os.path.realpath(__file__))
 sys.path.append(local_path)
